Decompose VM concern domains into modules
